Improve frontend styling and chat UI sequencing
Frontend styling:
- Add dark mode with OKLCH color system
- Add Comfortaa font via Google Fonts
- Update UI components (Button, Card, Input, Badge, Spinner) for dark mode
- Create Logo component with Dices icon
- Restructure routes with pathless _main layout for header/footer
- Make games list the home page (no redirect)
- Full-screen chat page without site header/footer

Chat UI improvements:
- Fix tool call sequencing (appear above response, stay visible after completion)
- Store tool calls as part of ChatMessage for proper ordering
- Show tool calls during streaming, then persist with message
- Fix markdown rendering to handle React children properly
- Add cursor-pointer to buttons
- Fix header margin/padding issues on chat page

Backend:
- Update chat system prompt with GameGame knowledge section
- Set temperature to 1.0 for GPT-5 compatibility
- Add GitHub URL and question type handling

+++ backend/src/gamegame/cli/maintenance.py
@@ -166,7 +166,7 @@ def storage_stats():
     asyncio.run(_stats())
 
 
-def _format_size(size: int) -> str:
+def _format_size(size: float) -> str:
     """Format bytes as human-readable size."""
     for unit in ["B", "KB", "MB", "GB", "TB"]:
         if size < 1024:
